Update 7.1.1.Final-install-on-rhel65.sh
Planning to automate jboss-7.1.1-Final, which is pretty much moved into stable territory as far as jboss releases are concerned. Not really personally used jboss but set it up with Mavern for current project manager on a rhel6.5. Looks like it is working well since he has not been back to me see for anything on that score. Following same steps and just setting it up with no application attached