Update Install CentOS 6 Server, join to AD for Centralized Authentication, and Join to in-house Spacewalk Server for Patching
This kictstart file will serve as a basis for a quick preconfigured install of a CentOS 6.5 or other Rhel-based distro by popping you rhel-based install media into your server (baremetal or virtual and typing in the reference command:
linux ks=nfs:192.168.1.23:/ks.cfg ip=dhcp  
and it will install your server, join it to Active Directory, and to the spacewalk or RHN Satellite server for updates and patching with no further intervention but for logging in with your domain credentials.

Place this kickstart file in the same directory as your full iso folder and do a mkisofs ... if you prefer cdrom-based kickstarting. This is actually my preferred method since nfs sometimes acts up. To kickstart from a cdrom issue the command and your server will be installed with no further interaction but for signing in:
linux ks=cdrom:/ks.cfg